Remove kiwipy/rmq dependencies of process module
