+++ src/plumpy/exceptions.py
@@ -39,5 +39,6 @@ class PersistenceError(Exception):
 class ClosedError(Exception):
     """Raised when an mutable operation is attempted on a closed process"""
 
+
 class TaskRejectedError(Exception):
-    """ A task was rejected by the coordinacor"""
+    """A task was rejected by the coordinacor"""
